Make the bazel command configurable.
It's not uncommon to have bazel in a custom location, or even to use a wrapper
for some projects, so we make it possible to set the command (defaulting to
"bazel").

This also introduces a new configuration group (for the new variable), and
collects both configuration groups under a common parent. And finally gives a
couple more discoverable names (e.g., `bazel-workspace-root`), since those are
functions likely to be called by consumers of this package.